+++ src/components/cameracontrols/captureAndTorch.tsx
